Create topTaxaPrevalence.R
This function takes in a phyloseq object (ps), an integer (n), and a string (""). ps will be used to calculate the top n most prevalent taxa across all samples. The function will return a bar graph displaying the top n lowestLevel's and the percent of samples in which they are present.