Update foodseqSetup.R
Adds section at the beginning that initializes a sample data frame in the case that the phyloseq object fed into the function doesn't have any metadata yet. It initalizes with the sample names as row names and adds the column, `Sample_ID` with the same values as the row names.