Update pcaPlot.R
Adds option for `customGradient` and `mid` in function call. This adds the ability to use a custom color palette for numerical variables (e.g., pMR, age) that has a `high`, `mid`, and `low` color assignment. `mid` allows you to set the midpoint of the gradient (i.e, mean, median, middle)