Create elbowPC.R
Code for elbow method to select relevant PC's. The goal of the code is to find the point on the scree curve that is the farthest from the straight line connecting the first and last points. This is the elbow. All PC's after this would be considered irrelevant. 

This is just one method for finding relevant PC's. 

Requires the output of pcaPlot() 

Takes in the output object from pcaPlot() and uses the scree table to extract the eigenvalues from PC generation.